tests/wpt/web-platform-tests/workers/Worker_script_mimetype.htm

<!DOCTYPE html>
<title> Worker constructor with script inside text file </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(function(t) {
  var worker = new Worker('./support/WorkerText.txt');
  worker.onmessage = t.step_func_done(function(e) {
    assert_equals(e.data, "Pass");
  });
  worker.postMessage("start");
});
</script>
